Reorganisation of the code into more organised chunks. Moving away from a more object-oriented organisation towards organising more by function or feature.
